// miniprogram/service/type.ts
export type weatherDataType = {
  air: {
    aqi: number
    aqi_level: number
    aqi_name: string
    co: string
    no2: string
    o3: string
    pm10: string
    "pm2.5": string
    so2: string
    update_time: string
  }
// ...
}

export type getWeatherReturnType = {
  data:weatherDataType
  message: string
  status: number
}